+++ design-system/button/extract-react-types/custom-theme-button-props.tsx
@@ -1,5 +1,7 @@
-import { OwnProps } from '../src/custom-theme-button/custom-theme-button-types';
+import { CustomThemeButtonOwnProps } from '../src/custom-theme-button/custom-theme-button-types';
 
-export default function CustomThemeButtonProps(props: OwnProps) {
+export default function CustomThemeButtonProps(
+  props: CustomThemeButtonOwnProps,
+) {
   return null;
 }