+++ design-system/button/src/entry-points/custom-theme-button.ts
@@ -4,4 +4,5 @@ export {
   ThemeProps,
   InteractionState,
   CustomThemeButtonProps,
+  CustomThemeButtonOwnProps,
 } from '../custom-theme-button/custom-theme-button-types';