+++ design-system/button/src/entry-points/types.ts
@@ -1,5 +1,5 @@
 // A seperate export which is only types
-export { Appearance, Spacing } from '../types';
+export { Appearance, Spacing, BaseOwnProps } from '../types';
 export { ButtonProps } from './standard-button';
 export { LoadingButtonProps } from './loading-button';
 export {
@@ -7,4 +7,5 @@ export {
   ThemeProps,
   InteractionState,
   CustomThemeButtonProps,
+  CustomThemeButtonOwnProps,
 } from './custom-theme-button';