+++ design-system/page-layout/src/controllers/types.ts
@@ -5,8 +5,12 @@ export type SkipLinkContextProps = {
 };
 
 export type SkipLinkData = {
+  /** id for the element that will be skipped to */
   id: string;
+  /** Text for the link that will appear in the skip link menu */
   skipLinkTitle: string;
+  /** Desired position in the skip link menu */
+  listIndex?: number;
 };
 
 export type SkipLinkI18n = {