+++ design-system/page-layout/src/index.ts
@@ -9,6 +9,7 @@ export {
   LeftPanel,
   Banner,
   TopNavigation,
+  useCustomSkipLink,
 } from './components';
 
 export {