+++ design-system/avatar-group/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/avatar-group
 
+## 8.0.9
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 8.0.8
 
 ### Patch Changes