+++ design-system/avatar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/avatar
 
+## 20.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 20.0.4
 
 ### Patch Changes