+++ design-system/badge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/badge
 
+## 14.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 14.0.4
 
 ### Patch Changes