+++ design-system/banner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/banner
 
+## 11.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.0.3
 
 ### Patch Changes