+++ design-system/blanket/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/blanket
 
+## 11.0.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.0.1
 
 ### Patch Changes