+++ design-system/datetime-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/datetime-picker
 
+## 10.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.0.4
 
 ### Patch Changes