+++ design-system/empty-state/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/empty-state
 
+## 7.1.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 7.1.2
 
 ### Patch Changes