+++ design-system/field-radio-group/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/field-radio-group
 
+## 8.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 8.0.2
 
 ### Patch Changes