+++ design-system/field-range/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/field-range
 
+## 9.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 9.0.2
 
 ### Patch Changes