+++ design-system/field-text/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/field-text
 
+## 11.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.0.2
 
 ### Patch Changes