+++ design-system/form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/form
 
+## 8.1.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 8.1.4
 
 ### Patch Changes