+++ design-system/global-navigation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/global-navigation
 
+## 10.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.0.3
 
 ### Patch Changes