+++ design-system/icon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/icon
 
+## 21.1.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 21.1.2
 
 ### Patch Changes