+++ design-system/inline-dialog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/inline-dialog
 
+## 13.0.7
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 13.0.6
 
 ### Patch Changes