+++ design-system/inline-edit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/inline-edit
 
+## 11.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.0.4
 
 ### Patch Changes