+++ design-system/navigation-next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/navigation-next
 
+## 9.0.8
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 9.0.7
 
 ### Patch Changes