+++ design-system/page-header/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/page-header
 
+## 10.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.0.3
 
 ### Patch Changes