+++ design-system/range/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/range
 
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 5.0.2
 
 ### Patch Changes