+++ design-system/section-message/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/section-message
 
+## 5.0.7
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 5.0.6
 
 ### Patch Changes