+++ design-system/table-tree/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/table-tree
 
+## 9.0.6
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 9.0.5
 
 ### Patch Changes