+++ design-system/tabs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tabs
 
+## 12.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 12.0.4
 
 ### Patch Changes