+++ design-system/textarea/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/textarea
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 3.0.3
 
 ### Patch Changes