+++ design-system/textfield/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/textfield
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 4.0.7
 
 ### Patch Changes