+++ editor/conversation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/conversation
 
+## 16.0.8
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 16.0.7
 
 ### Patch Changes