+++ media/media-editor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-editor
 
+## 38.1.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 38.1.0
 
 ### Minor Changes