+++ media/media-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-table
 
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 6.0.0
 
 ### Minor Changes