+++ media/media-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-ui
 
+## 14.1.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 14.1.1
 
 ### Patch Changes